+++ Java-Ring/src/main/java/org/project/object/armors/Armor.java
@@ -1,0 +1,52 @@
+package org.project.object.armors;
+import org.project.object.Object;
+
+// TODO: UPDATE IMPLEMENTATION
+public abstract class Armor implements Object {
+    private int defense;
+    private int maxDefense;
+    private int durability;
+    private int maxDurability;
+    private int weight;
+
+    private boolean isBroke;
+
+    public Armor(int defense, int durability) {
+        this.defense = defense;
+        this.durability = durability;
+    }
+
+    public void checkBreak() {
+        if (durability <= 0) {
+            isBroke = true;
+            defense = 0;
+        }
+    }
+
+    // TODO: (BONUS) UPDATE THE REPAIR METHOD
+    public void repair() {
+        isBroke = false;
+        defense = (int)(0.8*maxDefense);
+        durability = maxDurability;
+    }
+
+    public int getDefense() {
+        return defense;
+    }
+
+    public int getDurability() {
+        return durability;
+    }
+
+    public void reduceDurability(int durabilityCost) {
+        this.durability-=durabilityCost;
+    }
+
+    public int getWeight() {
+        return weight;
+    }
+
+    public boolean isBroke() {
+        return isBroke;
+    }
+}
